+++ CHANGELOG.md
@@ -1,11 +1,13 @@
 # *current*
 
+---
+
+# 0.0.5
+
 * Fixes crash on extra whitespace inside a pseudo selector's parentheses.
 * Adds sort function to the container class.
 * Enables the parser to pass its input through without transforming.
 * Iteration-safe `each` and `eachInside`.
-
----
 
 # 0.0.4